Updated main.py && added return values

+++ fish.py
@@ -23,19 +23,24 @@ def install_fish():
 
     print("Installing OhMyFish!")
     stdout, stderr, status = exec_command(f'fish ohmyfish.sh --noninteractive --yes')
+    if status != 0:
+        return False
     #print(stdout, stderr, status)
     stdout, stderr, status = exec_command(f'rm ohmyfish.sh')
+    if status != 0:
+        return False
     #print(stdout, stderr, status)
 
     print("Changing Shell")
     stdout, stderr, status = exec_command(f'chsh {user} -s /usr/bin/fish')
     print(stdout)
+    if status != 0:
+        return False
 
     print("Now set default shell on Konsole Profile")
+    return True
 
 
-
-#install_fish()
 
 def install_fisher():
     print('Downloading fisher')
@@ -47,14 +52,19 @@ def install_fisher():
         script.write(fisher_script)
 
     stdout, stderr, status = exec_command(['fish', '-c', 'source fisher.sh; fisher install jorgebucaran/fisher'], needs_split = False, as_root = False)
+    if status != 0:
+        return False
     print(stdout, stderr, status)
     
     exec_command('rm fisher.sh')
+    return True
     
 
 def install_fish_plugins():
     for plugin in fish_plugins:
         stdout, stderr, status = exec_command(f"fish -c 'fisher install {plugin}'", as_root = False)
         print(stdout, stderr, status)
+        if status != 0:
+            return False
+    return True
 
-install_fish_plugins()
